pyproject.toml
[project]
name = "offerpilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.135.3",
    "langchain>=1.2.15",
    "langgraph>=1.1.6",
    "loguru>=0.7.3",
    "path>=17.1.1",
    "psycopg>=3.2.0",
    "pydantic>=2.13.0",
    "pymupdf>=1.27.2.2",
    "pytest>=9.0.3",
    "rapidocr-onnxruntime",
    "ruamel-yaml>=0.19.1",
    "sqlalchemy>=2.0.0",
]
